requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
pytrends>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
ib_insync>=0.9.86
